Fix compilation errors: add InsertCallback, fix score tween types, remove unused fields
- Add Sequence.InsertCallback() to DOTween library
- Fix score count-up to use float getter/setter with int cast
- Remove unused _deathSequencePlaying and _pulseMin fields

+++ Assets/Scripts/UI/GameOverSequence.cs
@@ -78,15 +78,15 @@ public class GameOverSequence : MonoBehaviour
         // 0.6s: Score counts up from 0 to final
         if (_finalScoreText != null)
         {
-            int displayScore = 0;
+            float displayScore = 0f;
             float countDuration = Mathf.Min(finalScore * 0.03f, 1.5f);
             _sequence.Insert(0.6f,
                 DOTween.To(() => displayScore, v =>
                 {
                     displayScore = v;
-                    _finalScoreText.text = v.ToString();
+                    _finalScoreText.text = ((int)v).ToString();
                     _finalScoreText.color = Color.white;
-                }, finalScore, Mathf.Max(countDuration, 0.3f)).SetEase(Ease.OutQuad)
+                }, (float)finalScore, Mathf.Max(countDuration, 0.3f)).SetEase(Ease.OutQuad)
             );
         }
